chore: defaultGrid constant
