+++ packages/frontend/src/constants/index.ts
@@ -2,6 +2,14 @@ import { ButtonProps, MantineSize, PopoverProps } from '@mantine/core';
 
 export const MAX_PIVOTS = 3;
 
+export const defaultGrid = {
+    containLabel: true,
+    left: '5%', // small padding
+    right: '5%', // small padding
+    top: '70px', // pixels from top (makes room for legend)
+    bottom: '30px', // pixels from bottom (makes room for x-axis)
+} as const;
+
 export const FOOTER_HEIGHT = 80;
 export const FOOTER_MARGIN: MantineSize = 'lg';
