+++ packages/frontend/src/hooks/echarts/useEchartsCartesianConfig.ts
@@ -40,7 +40,7 @@ import moment from 'moment';
 import { useMemo } from 'react';
 import { isCartesianVisualizationConfig } from '../../components/LightdashVisualization/VisualizationConfigCartesian';
 import { useVisualizationContext } from '../../components/LightdashVisualization/VisualizationProvider';
-import { defaultGrid } from '../../components/VisualizationConfigs/ChartConfigPanel/Grid';
+import { defaultGrid } from '../../constants';
 import { EMPTY_X_AXIS } from '../cartesianChartConfig/useCartesianChartConfig';
 import getPlottedData from '../plottedData/getPlottedData';
